backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
anthropic==0.74.1
openai==2.8.1
pinecone==8.0.0
requests==2.32.5
python-docx==1.2.0
pypdf==6.4.0
python-pptx==1.0.2
tiktoken==0.12.0